% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{afile}
\alias{afile}
\title{afile}
\format{
\subsection{\code{afile}}{

A data frame with x rows and y columns:
\describe{
\item{id}{id}
}
}
}
\source{
Boyd
}
\usage{
afile
}
\description{
A subset of data from the ACS ...
}
\keyword{datasets}
