% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aranks}
\alias{aranks}
\title{aranks}
\format{
\subsection{\code{aranks}}{

A data frame with 10 rows and 3 columns:
\describe{
\item{ida}{ida}
\item{weighta}{weighta}
\item{ranka}{ranka}
}
}
}
\source{
Boyd
}
\usage{
aranks
}
\description{
Data to demonstrate and test the \code{matchrecs} function
}
\keyword{datasets}
