% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bfile}
\alias{bfile}
\title{bfile}
\format{
\subsection{\code{bfile}}{

A data frame with x rows and y columns:
\describe{
\item{id}{id}
}
}
}
\source{
Boyd
}
\usage{
bfile
}
\description{
A subset of data from the ACS ...
}
\keyword{datasets}
