% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{branks}
\alias{branks}
\title{branks}
\format{
\subsection{\code{branks}}{

A data frame with 5 rows and 3 columns:
\describe{
\item{idb}{idb}
\item{weightb}{weightb}
\item{rankb}{rankb}
}
}
}
\source{
Boyd
}
\usage{
branks
}
\description{
Data to demonstrate and test the \code{matchrecs} function
}
\keyword{datasets}
