% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expected_matchrecs}
\alias{expected_matchrecs}
\title{expected_matchrecs}
\format{
\subsection{\code{expected_matchrecs}}{

A data frame with 14 rows and 5 columns:
\describe{
\item{ida}{ida}
\item{idb}{idb}
\item{weight}{weight}
\item{ranka}{ranka}
\item{rankb}{rankb}
}
}
}
\source{
Boyd
}
\usage{
expected_matchrecs
}
\description{
Data to demonstrate and test the \code{matchrecs} function
}
\keyword{datasets}
