% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filematch-package.R
\docType{package}
\name{filematch-package}
\alias{filematch}
\alias{filematch-package}
\title{filematch: Tools For Matching Microdata Files}
\description{
This package has tools for matching microdata files.
}
\author{
\strong{Maintainer}: Don Boyd \email{donboyd5@gmail.com}

}
\keyword{internal}
