% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_cost_flow_helpers.R
\name{get_abfile}
\alias{get_abfile}
\title{Construct the AB file}
\usage{
get_abfile(arcs, nodes, flows, afile, bfile, idvar, wtvar, xvars, yvars, zvars)
}
\arguments{
\item{arcs}{dataframe with columns...}

\item{nodes}{dataframe with columns}

\item{flows}{vector...}

\item{afile}{dataframe with ...}

\item{bfile}{dataframe with ...}

\item{idvar}{character column name}

\item{wtvar}{character column name}

\item{xvars}{character vector of column names}

\item{yvars}{character vector of column names}

\item{zvars}{character vector of column names}
}
\value{
list with ...
}
\description{
Construct the AB file
}
\examples{
library(filematch)
data(afile)
}
