% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_cost_flow_helpers.R
\name{get_arcs}
\alias{get_arcs}
\title{Create arcs for a minimum cost flow problem}
\usage{
get_arcs(dbtoa, datob, nodes)
}
\arguments{
\item{dbtoa}{list returned from \code{FNN:get.knnx()}.}

\item{datob}{list returned from \code{FNN:get.knnx()}.}

\item{nodes}{dataframe with one row per row in afile and one row per row in bfile, with columns file, node, abrow.}
}
\value{
dataframe with arcs for a minimum cost flow problem.
}
\description{
Create arcs for a minimum cost flow problem
}
