% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_cost_flow_helpers.R
\name{get_k}
\alias{get_k}
\title{Determine k: number of nearest neighbors to find}
\usage{
get_k(nrows)
}
\arguments{
\item{nrows}{Integer scalar: number of records in the A (donor) file}
}
\value{
Integer scalar: number of nearest neighbors to find
}
\description{
Determine k: number of nearest neighbors to find
}
\examples{
get_k(1000)
get_k(nrow(afile))
}
