% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_cost_flow_helpers.R
\name{get_nodes}
\alias{get_nodes}
\title{Create nodes for a minimum cost flow problem}
\usage{
get_nodes(afile, bfile)
}
\arguments{
\item{afile}{dataframe with columns id, node, arow, weight, weightadj, iweight}

\item{bfile}{dataframe with columns id, node, arow, weight, weightadj, iweight}
}
\value{
dataframe with afile and bfile concatenated, with integer supply values for the minimum cost flow problem. The afile supplies will be negative because it will receive weights from the bfile. The bfile is the donor and its supplies will be positive. The supplies will sum to zero.
}
\description{
Creates a combined dataframe from afile and bfile. Its node column and supply column will be used as inputs to a minimum cost flow problem.
}
\examples{
library(filematch)
data(afile)
}
