% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{internal_matchrecs}
\alias{internal_matchrecs}
\title{Good approach}
\usage{
internal_matchrecs(
  adf,
  bdf,
  vida = "ida",
  vidb = "idb",
  vweighta = "weighta",
  vweightb = "weightb",
  vranka = "ranka",
  vrankb = "rankb"
)
}
\arguments{
\item{adf}{dataframe with A records}

\item{bdf}{dataframe with B records}

\item{vida}{string, A id variable name default "ida"}

\item{vidb}{string, b id variable name default "idb"}

\item{vweighta}{string, A weight variable name default "weighta"}

\item{vweightb}{string, B weight variable name default "weightb"}

\item{vranka}{string, A rank variable name default "ranka"}

\item{vrankb}{string, B rank variable name default "rankb"}
}
\value{
ab dataframe
}
\description{
Fast matching of records
}
