% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_cost_flow_helpers.R
\name{matchab}
\alias{matchab}
\title{Match A and B files}
\usage{
matchab(afile, bfile, idvar, wtvar, xvars, yvars, zvars, ka = NULL, kb = NULL)
}
\arguments{
\item{afile}{Dataframe ...}

\item{bfile}{Dataframe ...}

\item{idvar}{Character column name}

\item{wtvar}{Character column name}

\item{xvars}{Character vector of column names}

\item{yvars}{Character vector of column names}

\item{zvars}{Character vector of column names}

\item{ka}{Integer number of nearest neighbors to find in A file}

\item{kb}{Integer number of nearest neighbors to find in B file}
}
\value{
list with prep_list, mcfresult, and abfile
}
\description{
Match A and B files
}
\examples{
library(filematch)
data(afile)
}
