% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matchrecs}
\alias{matchrecs}
\title{Good approach}
\usage{
matchrecs(adf, bdf)
}
\arguments{
\item{adf}{a dataframe}

\item{bdf}{a dataframe}
}
\value{
ab dataframe
}
\description{
Fast matching of records
}
