% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_cost_flow_helpers.R
\name{prepab}
\alias{prepab}
\title{Prepare the afile and bfile}
\usage{
prepab(afile, bfile, idvar, wtvar, xvars, k = NULL)
}
\arguments{
\item{afile}{Dataframe...}

\item{bfile}{Dataframe...}

\item{idvar}{Character column name}

\item{wtvar}{Character column name}

\item{xvars}{Character vector of column names}

\item{k}{Integer number of nearest neighbors to find}
}
\value{
a list with nodes, arcs, and preptime
}
\description{
Prepare the afile and bfile
}
\examples{
library(filematch)
data(afile)
}
